Updated FlutterHue

+++ packages/flutter_hue/test/utils/color_converter_test.dart
@@ -199,13 +199,13 @@ void main() {
   group(
     "rgb to",
     () {
-      const List<int> rgb1 = [255, 0, 0];
-      const List<int> rgb2 = [0, 255, 0];
-      const List<int> rgb3 = [0, 0, 255];
-      const List<int> rgb4 = [128, 128, 128];
-      const List<int> rgb5 = [0, 0, 0];
-      const List<int> rgb6 = [255, 255, 255];
-      const List<int> rgb7 = [138, 72, 136];
+      const List<double> rgb1 = [255.0, 0.0, 0.0];
+      const List<double> rgb2 = [0.0, 255.0, 0.0];
+      const List<double> rgb3 = [0.0, 0.0, 255.0];
+      const List<double> rgb4 = [128.0, 128.0, 128.0];
+      const List<double> rgb5 = [0.0, 0.0, 0.0];
+      const List<double> rgb6 = [255.0, 255.0, 255.0];
+      const List<double> rgb7 = [138.0, 72.0, 136.0];
 
       test(
         "xy",
